refactor: move current_orthography_name to morphodict
