+++ CreeDictionary/morphodict/templatetags/morphodict_orth.py
@@ -6,6 +6,9 @@ from cree_sro_syllabics import sro2syllabics
 from django import template
 from django.utils.html import format_html
 
+# TODO: get rid of this import
+from utils import ORTHOGRAPHY_NAME
+
 from ..orthography import ORTHOGRAPHY
 
 register = template.Library()
@@ -84,3 +87,14 @@ def to_macrons(sro_circumflex: str) -> str:
     Transliterate SRO to macrons.
     """
     return sro_circumflex.translate(CIRCUMFLEX_TO_MACRON)
+
+
+@register.simple_tag(takes_context=True)
+def current_orthography_name(context):
+    """
+    Returns a pretty string of the currently active orthography.
+    The orthography is determined by the orth= cookie in the HTTP request.
+    """
+    # Determine the currently requested orthography:
+    request_orth = context.request.COOKIES.get("orth", ORTHOGRAPHY.default)
+    return ORTHOGRAPHY_NAME[request_orth]
